Assorted linters

Preconfigured linters, either as a collective list of linters or individually. "tidy" linters implement guidelines from the tidyverse style guide.

Usage

lint_starts_lowercase(x, ...)

lint_full_stop(x, ...)

lint_no_full_stop(x, ...)

lint_sentence_case(x, ...)

lint_title_case(x, ...)

tidy_title(x, ...)

tidy_param(x, name, description, ...)

tidy_return(x, ...)

tidy_seealso(x, ...)

tidy

Format

An object of class list of length 4.

Arguments

x

A roxygen2::roxy_tag() that is the subject of linting.

...

Additional arguments unused.

name, description

Used for roxygen2::roxy_tag()-specific linters.

Details

Refer to the individual roxygen2::roxy_tag() for the respective tag for argument details.

Functions

lint_starts_lowercase(): Lowercase start linting. (uses $raw for roxygen2::roxy_tag()s)

lint_full_stop(): Ends in a full stop. (uses $raw for roxygen2::roxy_tag()s)

lint_no_full_stop(): Does not end in a full stop. (uses $raw for roxygen2::roxy_tag()s)

lint_sentence_case(): Sentence case linting (uses $raw for roxygen2::roxy_tag()s)

lint_title_case(): Title case linting

tidy_title(): Tidy 'Sentence case' titles

tidy_param(): Tidy 'Sentence case' @param definitions

tidy_return(): Tidy 'Sentence case' @return definitions

tidy_seealso(): Tidy 'Sentence case' @seealso definitions

tidy: A list of all tidyverse style guide inspired linters